% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\docType{data}
\name{as.bool}
\alias{as.bool}
\title{Wrapper try_cast duckdb}
\format{
An object of class \code{call} of length 2.
}
\usage{
as.bool
}
\description{
Wrapper try_cast duckdb
}
\examples{
p90 |>
 mutate(
   across(
     starts_with(c("I")),
     ~ !!as.bool
   )
 )
}
\keyword{datasets}
