% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{attach_db}
\alias{attach_db}
\title{Attaches a duckdb database to a connection}
\usage{
attach_db(conn, dbdir, db, crypt = FALSE)
}
\arguments{
\item{conn}{duckdb connection, can be obtained via the get_conn function}

\item{dbdir}{Path to the duckdb file}

\item{db}{Alias name for the database}

\item{crypt}{If TRUE, passes the value of DUCKDB_ENCRYPTION_KEY as the key for the encrypted database}
}
\value{
Name of the duckdb database
}
\description{
Attaches a duckdb database to a connection
}
