% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{create_cube}
\alias{create_cube}
\title{Génère une requête SQL pour générer un cube duckdb}
\usage{
create_cube(table, dims, aggr)
}
\arguments{
\item{table}{Table duckdb/dbplyr}

\item{dims}{Vecteur caractère des dimensions}

\item{aggr}{Description des calculs des mesures du cube
dans une chaîne caractère}
}
\description{
Génère une requête SQL pour générer un cube duckdb
}
\examples{
table |>
 compute() |>
 create_cube(
   dims = c("dim1", "dim2"),
   aggr =
   "round(sum(n))::integer as eff,
    round(sum(n * is_present::int))::integer as eff_present"
 ) |>
 mutate(
   tx_present = round(eff_present / eff * 100, 2)
 ) -> tbl_cube
}
