% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{create_view}
\alias{create_view}
\title{Generates a SQL view from a dbplyr table}
\usage{
create_view(table, view_name = NULL)
}
\arguments{
\item{table}{duckdb/dbplyr table}

\item{view_name}{Name of the view; if NULL, a random name is generated}
}
\value{
Table corresponding to the view
}
\description{
Generates a SQL view from a dbplyr table
}
