% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_conn.R
\name{duckdb_conn}
\alias{duckdb_conn}
\title{Duckdb connection}
\description{
Singleton for duckdb connection
}
\examples{
# New connection
db = duckdb_conn$new()
# Show connection
print(db)
# Never use conn, create new connection
DBI::dbExecute(db$conn, "select 42 as value")
print(db)
# Reuse connection
DBI::dbExecute(db$conn, "select 44 as value")
print(db)
# Disconnect
db$disconnect()
# Show connection
print(db)
# After disconnect, create new connection
DBI::dbExecute(db$conn, "select 42 as value")
print(db)
# Disconnect
db$disconnect()
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ext}}{Extensions chargées}

\item{\code{dbdir}}{Nom de la base duckdb}

\item{\code{geometry}}{Nom de la géométrie, 'wk' (défaut) ou 'blob'}

\item{\code{bigint}}{Conversion des bigint, 'integer64' (défaut) ou 'numeric'}

\item{\code{conn}}{Connexion duckdb}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-duckdb_conn-new}{\code{duckdb_conn$new()}}
\item \href{#method-duckdb_conn-print}{\code{duckdb_conn$print()}}
\item \href{#method-duckdb_conn-is_connected}{\code{duckdb_conn$is_connected()}}
\item \href{#method-duckdb_conn-renew}{\code{duckdb_conn$renew()}}
\item \href{#method-duckdb_conn-refresh}{\code{duckdb_conn$refresh()}}
\item \href{#method-duckdb_conn-disconnect}{\code{duckdb_conn$disconnect()}}
\item \href{#method-duckdb_conn-clone}{\code{duckdb_conn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-new"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-new}{}}}
\subsection{Method \code{new()}}{
Crée un objet 'duckdb_conn'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$new(
  ext = "none",
  dbdir = ":memory:",
  geometry = c("wk", "blob"),
  bigint = c("integer64", "numeric")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ext}}{Extensions chargées 'core': httpfs+spatial, 'geo' core+h3
'all' geo+ead_stat. 'none' par défaut}

\item{\code{dbdir}}{Nom du la base duckdb, par défaut base stockée en mémoire}

\item{\code{geometry}}{Nom de la géométrie, 'wk' (défaut) ou 'blob'}

\item{\code{bigint}}{Conversion des bigint, 'integer64' (défaut) ou 'numeric'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Un objet 'duckdb_conn'
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-print"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-print}{}}}
\subsection{Method \code{print()}}{
Fonction 'print' dédiée
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Inutilisé}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-is_connected"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-is_connected}{}}}
\subsection{Method \code{is_connected()}}{
Connexion valide ?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$is_connected()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE/FALSE
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-renew"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-renew}{}}}
\subsection{Method \code{renew()}}{
Renouvelle la connexion sans changer de duckdb_conn
Attention: toutes les tables en mémoire sont perdues
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$renew()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-refresh}{}}}
\subsection{Method \code{refresh()}}{
Rafraichit les identifiants de connexion S3
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Déconnecte la connexion courante. Une nouvelle connexion sera créée
au prochain appel de '$conn'
Attention: toutes les tables en mémoire sont perdues
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-clone"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
