% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{get_conn}
\alias{get_conn}
\title{Renvoie une connexion duckdb}
\usage{
get_conn(dbdir = ":memory:", ext = "none", macro = TRUE, new = NULL, ...)
}
\arguments{
\item{dbdir}{Chemin vers le fichier duckdb, par défaut base stockée en mémoire}

\item{ext}{Indique les extensions chargées. 'core': spatial, 'geo' h3 et spatial,
'none' = pas d'extension. 'none' par défaut}

\item{macro}{Si TRUE, par défaut, les macros usuelles de la fonction add_macros sont chargées}

\item{new}{Obsolète, non pris en compte, présent pour raison de compatibilité
avec d'anciens programmes}
}
\description{
Renvoie une connexion duckdb
}
