% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{pg_attach}
\alias{pg_attach}
\title{Attaches a Postgres database to duckdb}
\usage{
pg_attach(conn, param, db = "pg", schema = "public")
}
\arguments{
\item{conn}{duckdb connection, can be obtained via the get_conn function}

\item{param}{list of parameters}

\item{db}{alias name for the database}

\item{schema}{Postgres schema name, defaults to 'public'}
}
\value{
dbExecute return code
}
\description{
Attaches a Postgres database to duckdb
}
