% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{refresh_secret}
\alias{refresh_secret}
\title{Creates a secret from S3 environment variables}
\usage{
refresh_secret(conn)
}
\arguments{
\item{conn}{duckdb connection}
}
\value{
duckdb return code
}
\description{
Creates a secret from S3 environment variables
}
