% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_csv}
\alias{tbl_csv}
\title{Returns a duckdb table from a CSV file including S3}
\usage{
tbl_csv(conn, path, ..., lower = FALSE, verbose = FALSE)
}
\arguments{
\item{conn}{duckdb connection}

\item{path}{Path to the CSV file}

\item{...}{Options to pass to the duckdb read_csv function}

\item{lower}{If TRUE, variable names are converted to lowercase}
}
\value{
duckdb table
}
\description{
Returns a duckdb table from a CSV file including S3
}
