% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_db}
\alias{tbl_db}
\title{Returns a list of tables from a duckdb database}
\usage{
tbl_db(conn, db, verbose = FALSE)
}
\arguments{
\item{conn}{duckdb connection, can be obtained via the get_conn function}

\item{db}{Name of the duckdb database}
}
\value{
List of tables
}
\description{
Returns a list of tables from a duckdb database
}
