% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_duckdb}
\alias{tbl_duckdb}
\title{Returns a list of tables from a duckdb database}
\usage{
tbl_duckdb(conn, name, db = NULL, crypt = FALSE)
}
\arguments{
\item{conn}{duckdb connection, can be obtained via the get_conn function}

\item{name}{Name of the duckdb database; ".duckdb" is appended automatically and the database is located in the
(s3perso)/duckdb directory}

\item{db}{Alias name for the database, defaults to path}

\item{crypt}{If TRUE, passes the value of DUCKDB_ENCRYPTION_KEY as the key for the encrypted database}
}
\value{
List of tables
}
\description{
Returns a list of tables from a duckdb database
}
