% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_list}
\alias{tbl_list}
\title{Returns a list of duckdb tables from a list
of paths to parquet files/directories}
\usage{
tbl_list(conn, paths, level = 0, lower = FALSE, verbose = FALSE)
}
\arguments{
\item{conn}{duckdb connection}

\item{paths}{List of paths to 'parquet' files}
}
\value{
liste List of duckdb tables
}
\description{
Returns a list of duckdb tables from a list
of paths to parquet files/directories
}
