% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_pqt}
\alias{tbl_pqt}
\title{Returns a duckdb table from a parquet file including S3}
\usage{
tbl_pqt(conn, path, level = 0, lower = FALSE, verbose = FALSE)
}
\arguments{
\item{conn}{duckdb connection}

\item{path}{Path to the parquet table/directory}

\item{level}{Number of levels for partitioned parquet files, default 0 if not partitioned}

\item{lower}{If TRUE, variable names are converted to lowercase}
}
\value{
List of duckdb tables
}
\description{
Returns a duckdb table from a parquet file including S3
}
