% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tempname}
\alias{tempname}
\title{Generate a temporary name}
\usage{
tempname(prefix = "temp")
}
\arguments{
\item{prefix}{Character prefix for the name}
}
\value{
Character string with temporary name
}
\description{
Generate a temporary name
}
\keyword{internal}
