% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_duckdb_parquet.R
\name{write_df_parquet}
\alias{write_df_parquet}
\title{Writes a parquet file from a data frame}
\usage{
write_df_parquet(df, conn, path, dir = NULL, keep = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{conn}{duckdb connection, e.g. db\$conn}

\item{path}{file name, defaults to the table name with .parquet extension}

\item{dir}{output directory, optionally an S3 path, defaults to current directory}

\item{keep}{If TRUE, returns the input data.frame, otherwise returns the table linked to the created parquet file. Defaults to FALSE.}
}
\value{
dplyr table linked to the created parquet file (see 'keep' parameter to retain the input data.frame)
}
\description{
Writes a parquet file from a data frame
}
