% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_duckdb_parquet.R
\name{write_duckdb_parquet}
\alias{write_duckdb_parquet}
\title{Écrit un fichier parquet depuis une table duckdb}
\usage{
write_duckdb_parquet(
  table,
  path = NULL,
  dir = NULL,
  compression_level = 3,
  partition = NULL,
  keep = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{table}{table duckdb}

\item{path}{nom du fichier ou répertoire si partition est renseigné, par défaut le nom de la table}

\item{dir}{répertoire de sortie, éventuellement s3}

\item{compression_level, }{entre 1 et 22, 3 par défaut}

\item{partition}{chaine de la forme "var1, var2" pour créer une partition, pas de partition par défaut}

\item{keep}{Si TRUE, renvoit la table fournie en entrée sinon la table liée au fichier parquet créé. FALSE par défaut.}

\item{verbose}{indique la requete SQL générée}
}
\value{
table dplyr liée au fichier parquet créé (voir paramètre 'keep' pour conserver la table en entrée)
}
\description{
Écrit un fichier parquet depuis une table duckdb
}
