% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_duckdb_parquet.R
\name{write_duckdb_parquet_sql}
\alias{write_duckdb_parquet_sql}
\title{Écrit un fichier parquet depuis une requête duckdb}
\usage{
write_duckdb_parquet_sql(
  conn,
  query,
  path,
  partition = "",
  compression_level = 3,
  verbose = FALSE
)
}
\arguments{
\item{conn}{Connexion duckdb, par exemple db$conn}

\item{query}{Requête duckdb}

\item{path}{nom du fichier ou répertoire si partition est renseigné}

\item{partition}{chaine de la forme "var1, var2" pour créer une partition, pas de partition par défaut}

\item{compression_level, }{entre 1 et 22, 3 par défaut}

\item{verbose}{indique la requete SQL générée}
}
\value{
Chemin du fichier parquet créé
}
\description{
Écrit un fichier parquet depuis une requête duckdb
}
