% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{attach_db}
\alias{attach_db}
\title{Attach une base duckdb à une connexion}
\usage{
attach_db(conn, dbdir, db, crypt = FALSE)
}
\arguments{
\item{conn}{Connexion duckdb, peut être obtenu par la fonction get_conn}

\item{dbdir}{Chemin vers le fichier duckdb}

\item{db}{Nom de l'alias de la base}

\item{crypt}{Si TRUE, passe la valeur de DUCKDB_ENCRYPTION_KEY comme clé de la base cryptée}
}
\value{
Nom de la base duckdb
}
\description{
Attach une base duckdb à une connexion
}
