% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{create_cube}
\alias{create_cube}
\title{Generates a SQL query to create a duckdb cube}
\usage{
create_cube(table, dims, aggr)
}
\arguments{
\item{table}{duckdb/dbplyr table}

\item{dims}{Character vector of dimensions}

\item{aggr}{Description of cube measure calculations as a character string}
}
\description{
Generates a SQL query to create a duckdb cube
}
\examples{
table |>
 compute() |>
 create_cube(
   dims = c("dim1", "dim2"),
   aggr =
   "round(sum(n))::integer as eff,
    round(sum(n * is_present::int))::integer as eff_present"
 ) |>
 mutate(
   tx_present = round(eff_present / eff * 100, 2)
 ) -> tbl_cube
}
