% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{create_view}
\alias{create_view}
\title{Génère une vue SQL à partir d'une table dbplyr}
\usage{
create_view(table, view_name = NULL)
}
\arguments{
\item{table}{Table duckdb/dbplyr}

\item{view_name}{Nom de la vue, si NULL, un nom aléatoire est généré}
}
\value{
Table correspondant à la vue
}
\description{
Génère une vue SQL à partir d'une table dbplyr
}
