% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_conn.R
\name{duckdb_conn}
\alias{duckdb_conn}
\title{Duckdb connection}
\description{
Singleton for duckdb connection
}
\examples{
# New connection
db = duckdb_conn$new()
# Show connection
print(db)
# Never use conn, create new connection
DBI::dbExecute(db$conn, "select 42 as value")
print(db)
# Reuse connection
DBI::dbExecute(db$conn, "select 44 as value")
print(db)
# Disconnect
db$disconnect()
# Show connection
print(db)
# After disconnect, create new connection
DBI::dbExecute(db$conn, "select 42 as value")
print(db)
# Disconnect
db$disconnect()
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ext}}{Loaded extensions}

\item{\code{dbdir}}{Name of the duckdb database}

\item{\code{geometry}}{Geometry format name, 'wk' (default) or 'blob'}

\item{\code{bigint}}{bigint conversion type, 'integer64' (default) or 'numeric'}

\item{\code{conn}}{duckdb connection}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-duckdb_conn-new}{\code{duckdb_conn$new()}}
\item \href{#method-duckdb_conn-print}{\code{duckdb_conn$print()}}
\item \href{#method-duckdb_conn-is_connected}{\code{duckdb_conn$is_connected()}}
\item \href{#method-duckdb_conn-renew}{\code{duckdb_conn$renew()}}
\item \href{#method-duckdb_conn-disconnect}{\code{duckdb_conn$disconnect()}}
\item \href{#method-duckdb_conn-clone}{\code{duckdb_conn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-new"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-new}{}}}
\subsection{Method \code{new()}}{
Creates a 'duckdb_conn' object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$new(
  ext = "none",
  dbdir = ":memory:",
  geometry = c("wk", "blob"),
  bigint = c("integer64", "numeric")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ext}}{Loaded extensions: 'core': httpfs+spatial, 'geo': core+h3,
'all': geo+ead_stat. Defaults to 'none'}

\item{\code{dbdir}}{Name of the duckdb database, default is an in-memory database}

\item{\code{geometry}}{Geometry format name, 'wk' (default) or 'blob'}

\item{\code{bigint}}{bigint conversion type, 'integer64' (default) or 'numeric'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 'duckdb_conn' object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-print"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-print}{}}}
\subsection{Method \code{print()}}{
Dedicated 'print' function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-is_connected"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-is_connected}{}}}
\subsection{Method \code{is_connected()}}{
Is the connection valid?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$is_connected()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE/FALSE
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-renew"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-renew}{}}}
\subsection{Method \code{renew()}}{
Renews the connection without changing the duckdb_conn object.
Warning: all in-memory tables are lost
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$renew()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnects the current connection. A new connection will be created
on the next call to '$conn'.
Warning: all in-memory tables are lost
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_conn-clone"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_conn-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{duckdb_conn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
