% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{get_conn}
\alias{get_conn}
\title{Returns a duckdb connection}
\usage{
get_conn(dbdir = ":memory:", ext = "none", macro = TRUE, new = NULL, ...)
}
\arguments{
\item{dbdir}{Path to the duckdb file, default is an in-memory database}

\item{ext}{Specifies loaded extensions. 'core': spatial, 'geo': h3 and spatial,
'none': no extension. Defaults to 'none'}

\item{macro}{If TRUE (default), standard macros from the add_macros function are loaded}

\item{new}{Deprecated, not used, present for backward compatibility
with older programs}
}
\description{
Returns a duckdb connection
}
