% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{refresh_secret}
\alias{refresh_secret}
\title{Crée un secret à partir des variables env S3}
\usage{
refresh_secret(conn)
}
\arguments{
\item{conn}{Connexion duckdb}
}
\value{
Code retour duckdb
}
\description{
Crée un secret à partir des variables env S3
}
