% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_csv}
\alias{tbl_csv}
\title{Renvoit une table duckdb depuis un fichier csv y.c S3}
\usage{
tbl_csv(conn, path, ..., lower = FALSE, verbose = FALSE)
}
\arguments{
\item{conn}{Connexion duckdb}

\item{path}{Chemin du fichier csv}

\item{...}{Options à passer à la fonction read_csv du duckdb}

\item{lower}{Si TRUE, les variables sont converties en minuscules}
}
\value{
Table duckdb
}
\description{
Renvoit une table duckdb depuis un fichier csv y.c S3
}
