% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_duckdb}
\alias{tbl_duckdb}
\title{Renvoie une liste de tables depuis une base duckdb}
\usage{
tbl_duckdb(conn, name, db = NULL, crypt = FALSE)
}
\arguments{
\item{conn}{Connexion duckdb, peut être obtenu par la fonction get_conn}

\item{name}{Nom de la base duckdb, ".duckdb" est ajouté automatiquement et la base est dans le répertoire
(s3perso)/duckdb}

\item{db}{Nom de l'alias de la base, par défaut path}

\item{crypt}{Si TRUE, passe la valeur de DUCKDB_ENCRYPTION_KEY comme clé de la base cryptée}
}
\value{
Liste de tables
}
\description{
Renvoie une liste de tables depuis une base duckdb
}
