% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_list}
\alias{tbl_list}
\title{Renvoit une liste de tables duckdb depuis une liste
de chemins vers des fichiers/répertoires parquet}
\usage{
tbl_list(conn, paths, level = 0, lower = FALSE, verbose = FALSE)
}
\arguments{
\item{conn}{Connexion duckdb}

\item{liste}{Liste de chemins vers les fichiers 'parquet'}
}
\value{
liste Liste de tables duckdb
}
\description{
Renvoit une liste de tables duckdb depuis une liste
de chemins vers des fichiers/répertoires parquet
}
