% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_pqt}
\alias{tbl_pqt}
\title{Renvoit une table duckdb depuis un fichier parquet y.c S3}
\usage{
tbl_pqt(conn, path, level = 0, lower = FALSE, verbose = FALSE)
}
\arguments{
\item{conn}{Connexion duckdb}

\item{path}{Chemin de la table/répertoire parquet}

\item{level}{Nombre de niveaux dans le cas de fichiers parquet partitionnés,
par défaut 0 si pas de partionnement}

\item{lower}{Si TRUE, les variables sont converties en minuscules}
}
\value{
Liste de tables duckdb
}
\description{
Renvoit une table duckdb depuis un fichier parquet y.c S3
}
