% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_duckdb_parquet.R
\name{write_df_parquet}
\alias{write_df_parquet}
\title{Écrit fichier parquet depuis une table duckdb}
\usage{
write_df_parquet(df, conn, path, dir = NULL, keep = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{conn}{Connexion duckdb, par exemple db$conn}

\item{path}{Nom du fichier, par défaut le nom de la table.parquet}

\item{dir}{Répertoire de sortie, éventuellement s3, par défaut répertoire courant}

\item{keep}{Si TRUE, renvoit le data.frame fourni en entrée sinon la table liée au fichier parquet créé. FALSE par défaut.}
}
\value{
table dplyr liée au fichier parquet créé (voir paramètre 'keep' pour conserver le data.frame en entrée)
}
\description{
Écrit fichier parquet depuis une table duckdb
}
