% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_duckdb_parquet.R
\name{write_duckdb_parquet}
\alias{write_duckdb_parquet}
\title{Writes a parquet file from a duckdb table}
\usage{
write_duckdb_parquet(
  table,
  path = NULL,
  dir = NULL,
  compression_level = 3,
  partition = NULL,
  keep = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{table}{duckdb table}

\item{path}{file name or directory if partition is specified, defaults to the table name}

\item{dir}{output directory, optionally an S3 path}

\item{compression_level}{between 1 and 22, default 3}

\item{partition}{string of the form "var1, var2" to create a partition, no partition by default}

\item{keep}{If TRUE, returns the input table, otherwise returns the table linked to the created parquet file. Defaults to FALSE.}

\item{verbose}{prints the generated SQL query}
}
\value{
dplyr table linked to the created parquet file (see 'keep' parameter to retain the input table)
}
\description{
Writes a parquet file from a duckdb table
}
