% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_duckdb_parquet.R
\name{write_duckdb_parquet_sql}
\alias{write_duckdb_parquet_sql}
\title{Writes a parquet file from a duckdb query}
\usage{
write_duckdb_parquet_sql(
  conn,
  query,
  path,
  partition = "",
  compression_level = 3,
  verbose = FALSE
)
}
\arguments{
\item{conn}{duckdb connection, e.g. db$conn}

\item{query}{duckdb query}

\item{path}{file name or directory if partition is specified}

\item{partition}{string of the form "var1, var2" to create a partition, no partition by default}

\item{compression_level}{between 1 and 22, default 3}

\item{verbose}{prints the generated SQL query}
}
\value{
Path of the created parquet file
}
\description{
Writes a parquet file from a duckdb query
}
